LayoutTests/fast/dom/setAttributeNS-prefix-and-null-namespace-expected.txt

Test that calling setAttributeNS() with a prefixed qualifiedName and null NS throws NAMESPACE_ERR.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElement('test').setAttributeNS(null, 'foo:bar', 'baz') threw exception Error: NamespaceError: DOM Exception 14.
PASS successfullyParsed is true

TEST COMPLETE
